Reorganize tree structure
There were too many nested directories. I reorganized the structure to
contain only the up-to-date code (so the Numba-powered one) in a more
standard tree structure as it is done in general for python packages.

Substantially the working code stays in a 1st level directory (called
pygalaxy in our case). The tests directory is in the root of the
repository